requirements.txt
TikTokLive>=6.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
